Port minecart simulation, the potato cannon and block-breaking items
The rail-following maths moved onto the cart's behaviour, tooltips are written into a consumer, and
Item#canAttackBlock became a cancellable event - so the items that used it answer one instead.

+++ src/main/java/com/simibubi/create/compat/thresholdSwitch/SophisticatedStorage.java
@@ -26,7 +26,7 @@ public class SophisticatedStorage implements ThresholdSwitchCompat {
 
 	@Override
 	public long getSpaceInSlot(ResourceHandler<ItemResource> inv, int slot) {
-		return (ItemHandlerHelpers.getSlotLimit((long) inv, slot) * ItemHandlerHelpers.getStackInSlot(inv, slot).getOrDefault(DataComponents.MAX_STACK_SIZE, 64)) / 64;
+		return (ItemHandlerHelpers.getSlotLimit(inv, slot) * ItemHandlerHelpers.getStackInSlot(inv, slot).getOrDefault(DataComponents.MAX_STACK_SIZE, 64)) / 64;
 	}
 
 }
